tests/integration/testdata/invoke/cdk/python/aws-lambda-function/requirements.txt
aws-cdk.core
aws-cdk.aws-lambda
aws-cdk.aws-iam